Fixed possible undefined device error
When DeviceRegistrationAttempted method was invoked, it was possible
that the device was not found from the devices list and caused undefined
error.

This could happen if gateway module was restarted after sending
registration message and backend would try to call the method before
device was inserted into devices array after restart etc.